webapp/requirements.txt
tensorflow-cpu==2.6.0
Pillow==8.4.0
fastapi==0.70.0
uvicorn==0.15.0
Jinja2==3.0.1
python-multipart==0.0.5
matplotlib==3.4.2
seaborn==0.11.1
scikit-learn==0.24.2
